eliminate some compiler warnings, update to newer libcircle

+++ src/dcp/cleanup.c
@@ -70,7 +70,7 @@ void DCOPY_do_cleanup(DCOPY_operation_t* op, \
     /*
      * Truncate file on last chunk (synchronous mode can write past end of file)
      */
-    int64_t bytes_written = (op->chunk + 1) * DCOPY_user_opts.chunk_size;
+    int64_t bytes_written = (int64_t)(op->chunk + 1) * (int64_t)DCOPY_user_opts.chunk_size;
     if(bytes_written >= op->file_size) {
         /* truncate file to appropriate size, to do this before
          * setting permissions in case file does not have write permission */
